requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
pytest>=7.4.0
opencv-python-headless>=4.9.0
numpy>=1.26.0

